Move pkg/idtools to daemon/internal/idtools
