+++ daemon/create_unix.go
@@ -11,10 +11,10 @@ import (
 	cerrdefs "github.com/containerd/errdefs"
 	"github.com/containerd/log"
 	"github.com/docker/docker/daemon/container"
+	"github.com/docker/docker/daemon/internal/idtools"
 	"github.com/docker/docker/daemon/pkg/oci"
 	volumemounts "github.com/docker/docker/daemon/volume/mounts"
 	volumeopts "github.com/docker/docker/daemon/volume/service/opts"
-	"github.com/docker/docker/pkg/idtools"
 	containertypes "github.com/moby/moby/api/types/container"
 	mounttypes "github.com/moby/moby/api/types/mount"
 	"github.com/opencontainers/selinux/go-selinux/label"
