+++ daemon/daemon.go
@@ -41,6 +41,7 @@ import (
 	"github.com/docker/docker/daemon/images"
 	"github.com/docker/docker/daemon/internal/distribution"
 	dmetadata "github.com/docker/docker/daemon/internal/distribution/metadata"
+	"github.com/docker/docker/daemon/internal/idtools"
 	"github.com/docker/docker/daemon/internal/image"
 	"github.com/docker/docker/daemon/internal/layer"
 	libcontainerdtypes "github.com/docker/docker/daemon/internal/libcontainerd/types"
@@ -62,7 +63,6 @@ import (
 	"github.com/docker/docker/dockerversion"
 	"github.com/docker/docker/pkg/authorization"
 	"github.com/docker/docker/pkg/fileutils"
-	"github.com/docker/docker/pkg/idtools"
 	"github.com/docker/docker/pkg/plugingetter"
 	"github.com/docker/docker/pkg/sysinfo"
 	"github.com/moby/buildkit/util/grpcerrors"
