+++ daemon/volume/mounts/mounts.go
@@ -7,9 +7,9 @@ import (
 	"syscall"
 
 	"github.com/containerd/log"
+	"github.com/docker/docker/daemon/internal/idtools"
 	"github.com/docker/docker/daemon/volume"
 	"github.com/docker/docker/daemon/volume/safepath"
-	"github.com/docker/docker/pkg/idtools"
 	"github.com/docker/docker/pkg/stringid"
 	mounttypes "github.com/moby/moby/api/types/mount"
 	"github.com/moby/sys/user"
