+++ daemon/volume/service/default_driver_stubs.go
@@ -3,8 +3,8 @@
 package service
 
 import (
+	"github.com/docker/docker/daemon/internal/idtools"
 	"github.com/docker/docker/daemon/volume/drivers"
-	"github.com/docker/docker/pkg/idtools"
 )
 
 func setupDefaultDriver(_ *drivers.Store, _ string, _ idtools.Identity) error { return nil }
