+++ daemon/volume/service/service.go
@@ -7,11 +7,11 @@ import (
 
 	"github.com/containerd/log"
 	"github.com/docker/docker/daemon/internal/directory"
+	"github.com/docker/docker/daemon/internal/idtools"
 	"github.com/docker/docker/daemon/volume"
 	"github.com/docker/docker/daemon/volume/drivers"
 	"github.com/docker/docker/daemon/volume/service/opts"
 	"github.com/docker/docker/errdefs"
-	"github.com/docker/docker/pkg/idtools"
 	"github.com/docker/docker/pkg/plugingetter"
 	"github.com/docker/docker/pkg/stringid"
 	"github.com/moby/moby/api/types/events"
