+++ daemon/volume/service/service_linux_test.go
@@ -6,12 +6,12 @@ import (
 	"path/filepath"
 	"testing"
 
+	"github.com/docker/docker/daemon/internal/idtools"
 	"github.com/docker/docker/daemon/volume"
 	volumedrivers "github.com/docker/docker/daemon/volume/drivers"
 	"github.com/docker/docker/daemon/volume/local"
 	"github.com/docker/docker/daemon/volume/service/opts"
 	"github.com/docker/docker/daemon/volume/testutils"
-	"github.com/docker/docker/pkg/idtools"
 	"gotest.tools/v3/assert"
 	is "gotest.tools/v3/assert/cmp"
 )
